# assets/docs/swagger-api-server/projects/panda-api-docs/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0.0"
  title: Athens (Panda) API
  description: API Documentation for Athens API Server. Please be inform some of our routes may not be resposive due to other socket authentication which Swagger hasn't. Our target host can be located on **localhost:8000**. If you have further questions and clarifications you may contact us immediately.
  termsOfService: 'http://dev2.neworientalclub88.com'
  contact:
    email: jhameil@neworientalclub88.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - Authorization
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /user/create:
    post:
      tags:
        - User
      description: Used to create user information
      summary: Please be informed that some of our routes may not respond due to WebSocket requirements/authentication.
      operationId: user_create
      consumes:
        - 'application/json'
      parameters:
        - in: body
          name: Request Body
          description: JSON user details creation
          required: true
          schema:
            $ref: "#/definitions/UserCreate_JSON"
      responses:
        "400":
          description: Error
          schema:
            type: object
            properties:
              status:
                type: integer
              message:
                type: string
        "401":
          description: This may return error due to duplicated fields
          schema:
            $ref: "#/definitions/UserCreate_ErrorRes401"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            type: string
  #User Update
...
definitions:
...
  UserCreate_JSON:
    properties:
      username:
        type: string
      password:
        type: string
      retypePassword:
        type: string
      email_address:
        type: string
      nickname:
        type: string
      full_name:
        type: string
  UserCreate_ErrorRes401:
    properties:
      message:
        type: object
        properties:
          saveDetails:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              modelIdentity:
                type: string
              attrNames:
                type: array
                items:
                  type: string
...